Reset AAID connection on TRANSACTION_FAILED
There have long been reports from the field that Keystore's attempts
to retrieve the application info for an app uid sometimes fail.

The bugreports that have any details all mention an
EX_TRANSACTION_FAILED (-129) failure, which apparently indicates a
low-level binder failure.

If that is the case, drop the binder connection on the assumption that
future requests on the same connection may also fail.  Any future
attempts to retrieve application info will then re-connect, hopefully
with a better chance of success.

Along the way, improve logging, fix a comment and correct an error arm.

Bug: 288040045
Bug: 291583874
Test: CtsKeystoreTestCases
Change-Id: Ifc73b560ec08f6907dd5d0f59a6506def3fde7d3