Deprecate get in ILegacyKeystore
This pull request initiates the deprecation of the
get method in the ILegacyKeystore interface. This
is the first step towards phasing out the legacy
keystore, which was intended to be disabled but
remains in use due to dependencies in the VPN and
WIFI code.

- The get method in ILegacyKeystore is marked as deprecated.
- Clear warnings are added to indicate that this method
  is no longer recommended for use and will be removed in the future.

The existing behavior of get is maintained for
now to avoid immediate disruptions to the VPN
and WIFI code.

Test: atest CtsKeystoreTestCases
Bug: 307460850
Change-Id: I912eed99d8e7cd35ed3a7a06096d911984a2eb41