Allow native applications with UID as ROOT to perform attestation.
Attestation application ID is collected for the applications having
package names but for the native applications due to unavailibility of
package name, it was failing to get AAID. For example key attestation
tests were failing in native tests - keystore2_client_tests.

Changes are made to allow caller with UID as AID_ROOT to perform
attestaion with indicating caller package info as AndroidSystem.

Bug: 349350267
Test: atest keystore2_client_tests
Change-Id: I5cb8189c526df74f73c53f240df9665ba5ad0b79