Update watchdog logs to backoff if logs stuck
At times the underlying thread can get stuck and keep
the watchdog occupied for longer peroids of time. While
there are a number of reasons for this we don't want to
clog the logs every second. So we increase the time between
logs to MAX_REPORT_TIMEOUT so that when threads are stuck
they do not continuously send logs.

Bug: 265778655
Test: atest libwatchdog_rs.test
Change-Id: Ib143ad75cdde80fbd8e3301e207aebe62ce2f018