Skip DE-critical system keys on clearNamespace
When clearing keys for a specific uid, the C++ Keystore code included a
hack that would skip over deleting keys owned by the system uid (1000)
and which were marked with a special flag (ag/2161344).

This hack was not transplanted when Keystore was re-written from C++ to
Rust for Android S, but there are now scenarios that implicitly rely on
this behaviour being present.  So reproduce the hack in the Rust code
that handles legacy keys.

Test: 11->15 upgrade tests
Bug: 377650521
Change-Id: I9ca0f0b31670eb419870a96df93c78c1cf17b795