Don't read through symlinks in fsverity_manifest_generator
We're trying to change the build system to remove the need for a
staging directory, where all the files are in the same location
as they are on the device. This means that fsverity_manifest_generator
will no longer be able to read through symlinks, because the files
that the symlink is trying to point to won't be at the same location.

Instead, copy the digests from the actual files pased on on-device
paths after all the files have been processed.

Bug: 394404628
Test: atest ComposHostTestCases
Change-Id: I37bde27797104499ffeaf280068d8a781b5ee660