Allow specifying relative paths in fsverity_manifest_generator
Right now, all files guarded by fsv_meta files have to be in a directory
with the same structure as they will have on-device. We'd like to
remove that restriction to make the build faster. Allow specifying
the on-device relative path for each input file to the fsverity manifest
manually, so that the script doesn't have to infer it from the file's
path on the host.

Bug: 394404628
Test: Manually
Change-Id: I5dd59cf2c224e2b3718b0fe4b495f37b0d17ff6d