Make fsverity_manifest_generator deterministic
Apparently python protobufs are not deterministic by default.
fsverity_manifest_generator generates artifacts that end up on the
device, so it's important for reproducible builds and build caching
that it's deterministic.

The python protobuf implementation has a method to configure it to
be deterministic by default globally, but that's more work, so start
with this for now.

Bug: 394404628
Test: Presubmits
Change-Id: Ie63237fa1279d6e4f068c9a34a6e1c0ff8d24cea